tests/wpt/web-platform-tests/fullscreen/api/document-fullscreen-enabled-active-document.html

<!DOCTYPE html>
<title>Document#fullscreenEnabled when the document is not the active document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe allowfullscreen></iframe>
<script>
var t = async_test();

onload = t.step_func(() => {
  var iframe = document.querySelector("iframe");
  var documentBeforeNav = iframe.contentDocument;

  iframe.onload = t.step_func_done(() => {
    assert_false(documentBeforeNav.fullscreenEnabled);
  });

  // Navigate the iframe
  window[0].location.href = '/common/blank.html';
});
</script>
